+++ .appveyor.yml
@@ -2,8 +2,8 @@ notifications:
   - provider: Slack
     incoming_webhook: https://hooks.slack.com/services/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
     on_build_success: false
-    on_build_failure: true
-    on_build_status_changed: true
+    on_build_failure: false
+    on_build_status_changed: false
   - provider: Email
     on_build_success: false
     on_build_failure: false
