+++ development/modules/auxiliary_release.py
@@ -79,7 +79,7 @@ def prepare_release_tests_1(constr):
     import numpy as np
 
     sys.path.insert(0, "../../../respy/tests")
-    from respy.tests.codes.random_model import generate_init
+    from respy.tests.random_model import generate_init
 
     # Prepare fresh subdirectories
     for which in ["old", "new"]:
@@ -167,7 +167,7 @@ def prepare_release_tests_2(constr):
     this function only.
     """
     sys.path.insert(0, "../../../respy/tests")
-    from respy.tests.codes.random_model import generate_init
+    from respy.tests.random_model import generate_init
 
     # Prepare fresh subdirectories
     for which in ["old", "new"]:
@@ -197,7 +197,7 @@ def prepare_release_tests_3(constr):
     import numpy as np
 
     sys.path.insert(0, "../../../respy/tests")
-    from respy.tests.codes.random_model import generate_init
+    from respy.tests.random_model import generate_init
 
     # Prepare fresh subdirectories
     for which in ["old", "new"]:
@@ -218,7 +218,7 @@ def prepare_release_tests_4(constr):
     this function only.
     """
     sys.path.insert(0, "../../../respy/tests")
-    from respy.tests.codes.random_model import generate_init
+    from respy.tests.random_model import generate_init
 
     # Prepare fresh subdirectories
     for which in ["old", "new"]:
@@ -256,7 +256,7 @@ def prepare_release_tests_5(constr):
     this function only.
     """
     sys.path.insert(0, "../../../respy/tests")
-    from respy.tests.codes.random_model import generate_init
+    from respy.tests.random_model import generate_init
 
     # Prepare fresh subdirectories
     for which in ["old", "new"]:
@@ -295,7 +295,7 @@ def prepare_release_tests_6(constr):
     import numpy as np
 
     sys.path.insert(0, "../../../respy/tests")
-    from respy.tests.codes.random_model import generate_init
+    from respy.tests.random_model import generate_init
 
     # Prepare fresh subdirectories
     for which in ["old", "new"]:
@@ -347,10 +347,9 @@ def prepare_release_tests_7(constr):
     """
     # This script is also imported (but not used) for the creation of the virtual environments.
     # Thus, the imports might not be valid when starting with a clean slate.
-    import numpy as np
 
     sys.path.insert(0, "../../../respy/tests")
-    from respy.tests.codes.random_model import generate_init
+    from respy.tests.random_model import generate_init
 
     # Prepare fresh subdirectories
     for which in ["old", "new"]:
@@ -394,10 +393,9 @@ def prepare_release_tests_8(constr):
     """
     # This script is also imported (but not used) for the creation of the virtual environments.
     # Thus, the imports might not be valid when starting with a clean slate.
-    import numpy as np
 
     sys.path.insert(0, "../../../respy/tests")
-    from respy.tests.codes.random_model import generate_init
+    from respy.tests.random_model import generate_init
 
     # Prepare fresh subdirectories
     for which in ["old", "new"]:
@@ -449,8 +447,8 @@ def prepare_release_tests_9(constr):
     import numpy as np
 
     sys.path.insert(0, "../../../respy/tests")
-    from respy.tests.codes.auxiliary import get_valid_shares
-    from respy.tests.codes.random_model import generate_init
+    from respy.tests.random_model import get_valid_shares
+    from respy.tests.random_model import generate_init
 
     # Prepare fresh subdirectories
     for which in ["old", "new"]:
@@ -501,10 +499,9 @@ def prepare_release_tests_10(constr):
 
     # This script is also imported (but not used) for the creation of the virtual environments.
     # Thus, the imports might not be valid when starting with a clean slate.
-    import numpy as np
 
     sys.path.insert(0, "../../../respy/tests")
-    from respy.tests.codes.random_model import generate_init
+    from respy.tests.random_model import generate_init
 
     # Prepare fresh subdirectories
     for which in ["old", "new"]:
@@ -559,7 +556,7 @@ def no_preparations_required(constr):
     this function only.
     """
     sys.path.insert(0, "../../../respy/tests")
-    from respy.tests.codes.random_model import generate_init
+    from respy.tests.random_model import generate_init
 
     # Prepare fresh subdirectories
     for which in ["old", "new"]:
@@ -589,7 +586,7 @@ def run_estimation(which):
     import respy
 
     sys.path.insert(0, os.path.dirname(respy.__file__) + "/tests")
-    from respy.tests.codes.auxiliary import simulate_observed
+    from respy.tests.random_model import simulate_observed
 
     init_dict = json.load(open("init_dict.respy.json", "r"))
