+++ respy/python/interface.py
@@ -312,8 +312,6 @@ def respy_interface(respy_obj, request, data=None):
             state_space,
             num_agents_sim,
             periods_draws_sims,
-            seed_sim,
-            file_sim,
             edu_spec,
             optim_paras,
             is_debug,
@@ -436,8 +434,6 @@ def minimal_simulation_interface(attr):
         state_space,
         attr["num_agents_sim"],
         periods_draws_sims,
-        attr["seed_sim"],
-        attr["file_sim"],
         attr["edu_spec"],
         attr["optim_paras"],
         attr["is_debug"],
