+++ respy/tests/random_model.py
@@ -2,6 +2,7 @@
 import json
 
 import numpy as np
+import pandas as pd
 from numpy.random import choice
 from numpy.random import randint
 from numpy.random import uniform
@@ -9,8 +10,13 @@ from numpy.random import uniform
 from respy.pre_processing.model_checking import check_model_attributes
 from respy.pre_processing.model_processing import _create_attribute_dictionary
 from respy.pre_processing.specification_helpers import csv_template
+from respy.python.interface import minimal_simulation_interface
+from respy.python.shared.shared_auxiliary import dist_class_attributes
+from respy.python.shared.shared_constants import DATA_FORMATS_SIM
+from respy.python.shared.shared_constants import DATA_LABELS_EST
+from respy.python.shared.shared_constants import DATA_LABELS_SIM
 from respy.python.shared.shared_constants import OPTIMIZERS
-from respy.tests.codes.auxiliary import get_valid_shares
+from respy.python.simulate.simulate_auxiliary import write_out
 
 
 def generate_random_model(
@@ -31,6 +37,8 @@ def generate_random_model(
     file_path : str
         save path for the output. The extensions .csv and .json are appended
         automatically.
+    myopic : bool
+        Indicator for myopic agents meaning the discount factor is set to zero.
 
     """
     # potential conversions from numpy._bool to python bool. Don't remove!
@@ -138,7 +146,7 @@ def generate_random_model(
     options["interpolation"]["points"] = randint(10, 100)
 
     for optimizer in OPTIMIZERS:
-        options[optimizer] = generate_optimizer_options(optimizer, params)
+        options[optimizer] = generate_optimizer_options(optimizer)
 
     attr = _create_attribute_dictionary(params, options)
     check_model_attributes(attr)
@@ -167,7 +175,7 @@ def _consolidate_bound_constraints(bound_constr):
     return constr
 
 
-def generate_optimizer_options(which, params_spec):
+def generate_optimizer_options(which):
     dict_ = {}
 
     if which == "SCIPY-BFGS":
@@ -193,3 +201,115 @@ def generate_optimizer_options(which, params_spec):
         raise NotImplementedError("The optimizer you requested is not implemented.")
 
     return dict_
+
+
+def get_valid_shares(num_groups):
+    """We simply need a valid request for the shares of types summing to one."""
+    shares = np.random.uniform(size=num_groups)
+    shares = shares / np.sum(shares)
+    shares = shares.tolist()
+    return shares
+
+
+def simulate_observed(respy_obj, is_missings=True):
+    """ This function adds two important features of observed datasests: (1) missing
+    observations and missing wage information.
+    """
+
+    def drop_agents_obs(agent):
+        """ We now determine the exact period from which onward the history is truncated
+        and cut the simulated dataset down to size.
+        """
+        start_truncation = np.random.choice(range(1, agent["Period"].max() + 2))
+        agent = agent[agent["Period"] < start_truncation]
+        return agent
+
+    seed_sim = dist_class_attributes(respy_obj, "seed_sim")
+
+    _, df = respy_obj.simulate()
+
+    # It is important to set the seed after the simulation call. Otherwise, the value of
+    # the seed differs due to the different implementations of the PYTHON and FORTRAN
+    # programs.
+    np.random.seed(seed_sim)
+
+    # We read in the baseline simulated dataset.
+    data_frame = pd.read_csv(
+        "data.respy.dat",
+        delim_whitespace=True,
+        header=0,
+        na_values=".",
+        dtype=DATA_FORMATS_SIM,
+        names=DATA_LABELS_SIM,
+    )
+
+    if is_missings:
+        # We truncate the histories of agents. This mimics the frequent empirical fact
+        # that we loose track of more and more agents over time.
+        data_subset = data_frame.groupby("Identifier").apply(drop_agents_obs)
+
+        # We also want to drop the some wage observations. Note that we might be dealing
+        # with a dataset where nobody is working anyway.
+        is_working = data_subset["Choice"].isin([1, 2])
+        num_drop_wages = int(np.sum(is_working) * np.random.uniform(high=0.5, size=1))
+        if num_drop_wages > 0:
+            indices = data_subset["Wage"][is_working].index
+            index_missing = np.random.choice(indices, num_drop_wages, False)
+            data_subset.loc[index_missing, "Wage"] = None
+        else:
+            pass
+    else:
+        data_subset = data_frame
+
+    # We can restrict the information to observed entities only.
+    data_subset = data_subset[DATA_LABELS_EST]
+    write_out(respy_obj, data_subset)
+
+    return respy_obj
+
+
+def minimal_simulate_observed(attr, is_missings=True):
+    """ This function adds two important features of observed datasests: (1) missing
+    observations and missing wage information.
+    """
+
+    def drop_agents_obs(agent):
+        """ We now determine the exact period from which onward the history is truncated
+        and cut the simulated dataset down to size.
+        """
+        start_truncation = np.random.choice(range(1, agent["Period"].max() + 2))
+        agent = agent[agent["Period"] < start_truncation]
+        return agent
+
+    state_space, df = minimal_simulation_interface(attr)
+
+    np.random.seed(attr["seed_sim"])
+
+    if is_missings:
+        # We truncate the histories of agents. This mimics the frequent empirical fact
+        # that we loose track of more and more agents over time.
+        data_subset = df.groupby("Identifier").apply(drop_agents_obs)
+
+        # We also want to drop the some wage observations. Note that we might be dealing
+        # with a dataset where nobody is working anyway.
+        is_working = data_subset["Choice"].isin([1, 2])
+        num_drop_wages = int(np.sum(is_working) * np.random.uniform(high=0.5, size=1))
+        if num_drop_wages > 0:
+            indices = data_subset["Wage"][is_working].index
+            index_missing = np.random.choice(indices, num_drop_wages, False)
+            data_subset.loc[index_missing, "Wage"] = None
+        else:
+            pass
+    else:
+        data_subset = df
+
+    # We can restrict the information to observed entities only.
+    data_subset = data_subset[DATA_LABELS_EST]
+
+    # We maintain several versions of the file.
+    with open(attr["file_sim"] + ".respy.dat", "w") as file_:
+        data_subset.to_string(file_, index=False, header=True, na_rep=".")
+    data_subset.to_pickle(attr["file_sim"] + ".respy.pkl")
+    data_subset.Wage = data_subset.Wage.round(6)
+
+    return data_subset
